add support for doubles

+++ Calculator24/Program.cs
@@ -1,9 +1,9 @@
-﻿int Add(int a, int b)
+﻿double Add(double a, double b)
 {
     return a + b;
 }
 
-int Subtract(int a, int b)
+double Subtract(double a, double b)
 {
     return a - b;
 }
